task 6 corr2

+++ 0x1A-hash_tables/6-hash_table_delete.c
@@ -35,17 +35,18 @@ void hash_table_delete(hash_table_t *ht)
 	unsigned long int i;
 	hash_node_t *temp;
 
+	if (ht == NULL)
+		return;
 	for (i = 0; i < ht->size; i++)
 	{
 		if (ht->array[i] != NULL)
 		{
-			while (ht->array[i]->next != NULL)
+			while (ht->array[i] != NULL)
 			{
 				temp = ht->array[i]->next;
 				delete_item(ht->array[i]);
 				ht->array[i] = temp;
 			}
-			delete_item(ht->array[i]);
 		}
 	}
 	free(ht->array);
